requirements.txt
telebot
puowm
time
pyowm.exceptions